Add sign up, login, auth context, private route component, login and signup redirects
