--- src/components/TestHeader.jsx
@@ -1,6 +1,0 @@
-import { useAuth } from '../context/AuthContext';
-
-export default function TestHeader() {
-  const { user } = useAuth();
-  return <header>{user.username}</header>;
-}
